Add ability to interpolate any file content or env value
