+++ pkg/controller/appdefinition/files_test.go
@@ -14,3 +14,7 @@ func TestFileModes(t *testing.T) {
 func TestFileModesBug(t *testing.T) {
 	tester.DefaultTest(t, scheme.Scheme, "testdata/files-bug", DeploySpec)
 }
+
+func TestInterpolation(t *testing.T) {
+	tester.DefaultTest(t, scheme.Scheme, "testdata/interpolation", DeploySpec)
+}
