+++ pkg/controller/appdefinition/testdata/interpolation/existing.yaml.d/secret.yaml
@@ -1,0 +1,8 @@
+kind: Secret
+apiVersion: v1
+metadata:
+  name: sec-1
+  namespace: app-created-namespace
+data:
+  key1: "dmFsdWUx"
+  key2: "dmFsdWUy"
