+++ pkg/controller/appdefinition/testdata/interpolation/expected.yaml.d/pdb.yaml
@@ -1,0 +1,18 @@
+kind: PodDisruptionBudget
+apiVersion: policy/v1
+metadata:
+  name: oneimage
+  namespace: app-created-namespace
+  labels:
+    "acorn.io/app-namespace": "app-namespace"
+    "acorn.io/app-name": "app-name"
+    "acorn.io/container-name": "oneimage"
+    "acorn.io/managed": "true"
+spec:
+  selector:
+    matchLabels:
+      "acorn.io/app-namespace": "app-namespace"
+      "acorn.io/app-name": "app-name"
+      "acorn.io/container-name": "oneimage"
+      "acorn.io/managed": "true"
+  maxUnavailable: 25%
