+++ pkg/controller/appdefinition/testdata/interpolation/expected.yaml.d/serviceaccount.yaml
@@ -1,0 +1,10 @@
+kind: ServiceAccount
+apiVersion: v1
+metadata:
+  name: oneimage
+  namespace: app-created-namespace
+  labels:
+    acorn.io/app-name: app-name
+    acorn.io/app-namespace: app-namespace
+    acorn.io/managed: "true"
+    acorn.io/container-name: oneimage
