+++ pkg/controller/secrets/testdata/secret-encrypted/existing.yaml.d/secret.yaml
