+++ pkg/controller/secrets/testdata/secret-encrypted/expected.yaml.d/appinstance.yaml
