+++ pkg/controller/secrets/testdata/secret-encrypted/input.yaml
