+++ pkg/controller/secrets/testdata/secret-image/expected.yaml.d/appinstance.yaml
