+++ pkg/controller/secrets/testdata/secret-image/input.yaml
